+++ lib/options/atlas_to_sprites_options.d.ts
@@ -1,0 +1,8 @@
+/**
+ * The options that can be passed to the `atlasToSprites` method.
+ */
+export interface AtlasToSpritesOptions {
+    jsonPropertyPath?: string;
+    crossOrigin?: string;
+    download?: boolean;
+}