+++ lib/options/spritesheet_to_sprites_options.d.ts
@@ -1,0 +1,8 @@
+/**
+ * The options that can be passed to the `spritesheetToSprites` method.
+ */
+export interface SpritesheetToSpritesOptions {
+    name?: string;
+    crossOrigin?: string;
+    download?: boolean;
+}