+++ lib/utils.d.ts
@@ -1,0 +1,7 @@
+import { Sprite } from './interfaces/sprite';
+/**
+ * Downloads the sprites parsed from the spritesheet.
+ *
+ * @param {Array<Sprite>} sprites The sprites parsed from the spritesheet.
+ */
+export declare function downloadSprites(sprites: Array<Sprite>): void;