+++ test/tests/atlas_to_sprites_xml.test.js
@@ -2,54 +2,56 @@
 
 import { atlasToSprites, loadSpritesheet, loadXML } from '../gingerale.js';
 
-describe('Getting individual sprites from an atlas', () => {
-    it('should load an atlas using the resource-loader and parse it into individual sprites', async function () {
+describe('Getting individual sprites from an atlas', function () {
+    it('should load an atlas using the resource-loader and parse it into individual sprites', function () {
         const loader = new Loader();
 
         loader
             .add('foliagePackSpritesheet', './assets/atlas/xml/foilage_pack/foliagePack_default.png')
             .add('foliagePackXML', './assets/atlas/xml/foilage_pack/foliagePack_default.xml')
-            .load(async function (_, resources) {
-                const sprites = await atlasToSprites(resources.foliagePackSpritesheet.data, resources.foliagePackXML.data).catch(err => { throw err; });
+            .load(function (_, resources) {
+                const sprites = atlasToSprites(resources.foliagePackSpritesheet.data, resources.foliagePackXML.data);
 
                 chai.expect(sprites.length).to.equal(106);;
             });
     });
 
-    it('should set the correct width and heights for sprites', async function () {
+    it('should set the correct width and heights for sprites', function () {
         const loader = new Loader();
 
         loader
             .add('foliagePackSpritesheet', './assets/atlas/xml/foilage_pack/foliagePack_default.png')
             .add('foliagePackXML', './assets/atlas/xml/foilage_pack/foliagePack_default.xml')
-            .load(async function (_, resources) {
-                const sprites = await atlasToSprites(resources.foliagePackSpritesheet.data, resources.foliagePackXML.data).catch(err => { throw err; });
+            .load(function (_, resources) {
+                const sprites = atlasToSprites(resources.foliagePackSpritesheet.data, resources.foliagePackXML.data);
 
-                chai.expect(sprites[4].image.width).to.equal(102);
-                chai.expect(sprites[4].image.height).to.equal(207);
+                setTimeout(function () {
+                    chai.expect(sprites[4].image.width).to.equal(102);
+                    chai.expect(sprites[4].image.height).to.equal(207);
 
-                chai.expect(sprites[37].image.width).to.equal(122);
-                chai.expect(sprites[37].image.height).to.equal(172);
+                    chai.expect(sprites[37].image.width).to.equal(122);
+                    chai.expect(sprites[37].image.height).to.equal(172);
 
-                chai.expect(sprites[90].image.width).to.equal(64);
-                chai.expect(sprites[90].image.height).to.equal(64);
+                    chai.expect(sprites[90].image.width).to.equal(64);
+                    chai.expect(sprites[90].image.height).to.equal(64);
 
-                chai.expect(sprites[100].image.width).to.equal(64);
-                chai.expect(sprites[100].image.height).to.equal(64);
+                    chai.expect(sprites[100].image.width).to.equal(64);
+                    chai.expect(sprites[100].image.height).to.equal(64);
 
-                chai.expect(sprites[105].image.width).to.equal(15);
-                chai.expect(sprites[105].image.height).to.equal(15);
+                    chai.expect(sprites[105].image.width).to.equal(15);
+                    chai.expect(sprites[105].image.height).to.equal(15);
+                }, 1000);
             });
     });
 
-    it('should set the names for the sprites', async function () {
+    it('should set the names for the sprites', function () {
         const loader = new Loader();
 
         loader
             .add('foliagePackSpritesheet', './assets/atlas/xml/foilage_pack/foliagePack_default.png')
             .add('foliagePackXML', './assets/atlas/xml/foilage_pack/foliagePack_default.xml')
-            .load(async function (_, resources) {
-                const sprites = await atlasToSprites(resources.foliagePackSpritesheet.data, resources.foliagePackXML.data).catch(err => { throw err; });
+            .load(function (_, resources) {
+                const sprites = atlasToSprites(resources.foliagePackSpritesheet.data, resources.foliagePackXML.data);
 
                 chai.expect(sprites[4].name).to.equal('foliagePack_005.png');
                 chai.expect(sprites[37].name).to.equal('foliagePack_038.png');
@@ -59,14 +61,14 @@ describe('Getting individual sprites from an atlas', () => {
             });
     });
 
-    it('should set the (x, y) for the sprites', async function () {
+    it('should set the (x, y) for the sprites', function () {
         const loader = new Loader();
 
         loader
             .add('foliagePackSpritesheet', './assets/atlas/xml/foilage_pack/foliagePack_default.png')
             .add('foliagePackXML', './assets/atlas/xml/foilage_pack/foliagePack_default.xml')
-            .load(async function (_, resources) {
-                const sprites = await atlasToSprites(resources.foliagePackSpritesheet.data, resources.foliagePackXML.data).catch(err => { throw err; });
+            .load(function (_, resources) {
+                const sprites = atlasToSprites(resources.foliagePackSpritesheet.data, resources.foliagePackXML.data);
 
                 chai.expect(sprites[4].x).to.equal(483);
                 chai.expect(sprites[4].y).to.equal(588);
@@ -89,7 +91,7 @@ describe('Getting individual sprites from an atlas', () => {
         const spritesheet = await loadSpritesheet('./assets/atlas/xml/foilage_pack/foliagePack_default.png').catch(err => { throw err; });
         const xml = await loadXML('./assets/atlas/xml/foilage_pack/foliagePack_default.xml').catch(err => { throw err; });
 
-        const sprites = await atlasToSprites(spritesheet, xml).catch(err => { throw err; });
+        const sprites = atlasToSprites(spritesheet, xml);
 
         chai.expect(sprites.length).to.equal(106);
     });